requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
discord.py==1.5.1
idna==2.10
multidict==4.7.6
PyYAML==5.3.1
requests==2.25.0
urllib3==1.26.2
yarl==1.5.1
